Remove impressionist gem
To avoid the following warning:
```
DEPRECATION WARNING: Initialization autoloaded the constants SoftDeletable and ImpressionistController.

Being able to do this is deprecated. Autoloading during initialization is going
to be an error condition in future versions of Rails.

Reloading does not reboot the application, and therefore code executed during
initialization does not run again. So, if you reload SoftDeletable, for example,
the expected changes won't be reflected in that stale Module object.

These autoloaded constants have been unloaded.

Please, check the "Autoloading and Reloading Constants" guide for solutions.
```