Rails.application.reloader.to_prepare で囲む
以下のワーニングを鎮めるため。

```
DEPRECATION WARNING: Initialization autoloaded the constant BatchDestroyable.

Being able to do this is deprecated. Autoloading during initialization is going
to be an error condition in future versions of Rails.

Reloading does not reboot the application, and therefore code executed during
initialization does not run again. So, if you reload BatchDestroyable, for example,
the expected changes won't be reflected in that stale Module object.

This autoloaded constant has been unloaded.

In order to autoload safely at boot time, please wrap your code in a reloader
callback this way:

    Rails.application.reloader.to_prepare do
      # Autoload classes and modules needed at boot time here.
    end

That block runs when the application boots, and every time there is a reload.
For historical reasons, it may run twice, so it has to be idempotent.

Check the "Autoloading and Reloading Constants" guide to learn more about how
Rails autoloads and reloads.
 (called from <top (required)> at /workspace/config/environment.rb:7)
```